use top-level markdown headings + minor edits
This commit makes some stylistic changes to the README file. Previously,
the README headings used #### for headings, but because this is markdown
format, we should probably use # to designate top-level headings. This
also formats sub-headings with ## to indicate that they are part of the
top-level heading section.

This commit also fixes the link to the paper. Previously, http was
included twice.